[ICP GPU] Intermediate functions (#18)
* gpu icp: compilation matrix print

* gpu icp: compilable version of get_correspondances

* gpu icp: add mat transpose

* gpu icp: update transpose

* gpu icp: fixing compilation

* gpu icp: movie transpose

* cpu icp: checking correspondance indices with array pointers

* cpu fixing all array functions

* gpu icp: fixed functions cross correlation and correspondance

* gpu icp: fixing all functions and adding transpose test

* gpu icp: transpose test

* gpu icp: fonctionnal transpose tested

* gpu icp : fixed usage of std:: in cuda

* gpu icp: updated gpu cross function

* gpu icp: adding transpose wrapper

* icp gpu: fixing wrapper transpose

* gpu icp :adding call to kernel in cross variance

* gpu icp: in progress main call for testing

* gpu icp: added test to main

* icp gpu : checked kernel call for transpose good

* icp gpu: main test done

* src gpu: added call to kernel mult

* icp gpu: test call matrixmult kernel

* gpu icp: fixing call to mult kernel

* GPU Mult test fix

+++ src/libgpualg/mult.cu
@@ -4,18 +4,18 @@
 
 #define Tile_size 2
 
-__global__ void matrixMultiplyShared(float* A, float* B, float* C,
+__global__ void matrixMultiplyShared(double* A, double* B, double* C,
     int matARows, int matAColumns,
     int matBRows, int matBColumns,
     int matCRows, int matCColumns)
 {
     // Tile size to store elements in shared memory
-    __shared__ float sA[Tile_size][Tile_size]; 
-    __shared__ float sB[Tile_size][Tile_size];
+    __shared__ double sA[Tile_size][Tile_size]; 
+    __shared__ double sB[Tile_size][Tile_size];
 
     int Row = blockDim.y * blockIdx.y + threadIdx.y; //To generate ids of threads.
     int Col = blockDim.x * blockIdx.x + threadIdx.x;
-    float Cvalue = 0.0;
+    double Cvalue = 0.0;
     sA[threadIdx.y][threadIdx.x] = 0.0;
     sB[threadIdx.y][threadIdx.x] = 0.0;
 
@@ -52,7 +52,7 @@ __global__ void matrixMultiplyShared(float* A, float* B, float* C,
     }
 }
 
-void matrixMultiplication(float* A, float* B, float* C, 
+void matrixMultiplication(double* A, double* B, double* C, 
     int numARows, int numAColumns,
     int numBRows, int numBColumns,
     int numCRows, int numCColumns) {
